MortgageBroker/requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.19.0
airtable==0.4.8
airtable-python-wrapper==0.15.3
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.0
argilla==1.12.0
asgiref==3.6.0
asn1crypto==1.5.1
async-timeout==4.0.2
attrs==22.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.11.2
blis==0.7.9
blobfile==2.0.1
boto3==1.26.74
botocore==1.29.74
bs4==0.0.1
cachetools==5.3.1
camelot-py==0.11.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
cmake==3.26.3
colorama==0.4.6
commonmark==0.9.1
comtypes==1.1.14
confection==0.0.4
contourpy==1.0.6
cryptography==40.0.2
cssselect==1.2.0
cycler==0.11.0
cymem==2.0.7
dataclasses-json==0.5.8
defusedxml==0.7.1
Deprecated==1.2.14
distlib==0.3.6
Django==4.1.7
django-braces==1.15.0
django-cors-headers==3.13.0
django-db-tools==0.0.1
django-mysql==4.8.0
django-oauth-toolkit==2.3.0
django-rest==0.8.7
django-rest-framework-social-oauth2==1.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.3.0
docx2txt==0.8
drf-social-oauth2==2.1.3
ecdsa==0.18.0
email-validator==2.0.0.post2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
et-xmlfile==1.1.0
evals==1.0.3.post1
exceptiongroup==1.1.2
fastapi==0.95.1
fastapi-users==11.0.0
fastapi-users-db-sqlalchemy==5.0.0
feedfinder2==0.0.4
feedparser==6.0.10
ffmpy==0.3.0
filelock==3.9.0
filetype==1.2.0
fire==0.5.0
Flask==2.2.3
fonttools==4.38.0
fpdf==1.7.2
frozenlist==1.3.3
fsspec==2023.6.0
ghostscript==0.7
google-api-core==2.11.0
google-auth==2.20.0
google-cloud-documentai==2.15.0
googleapis-common-protos==1.59.1
googletrans==3.1.0a0
gpt==0.2
gradio==3.27.0
gradio_client==0.2.6
greenlet==2.0.1
grpcio==1.54.2
grpcio-status==1.54.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
html5lib==1.1
htmlmin==0.1.12
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.15.1
hyperframe==5.2.0
idna==2.10
ImageHash==4.3.1
inexactsearch==1.0.2
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.1
jsonschema==4.17.3
jsonschema-specifications==2023.6.1
jwcrypto==1.5.0
kiwisolver==1.4.4
langchain==0.0.188
langcodes==3.3.0
langdetect==1.0.9
linkify-it-py==2.0.2
llama-index==0.6.15
locationtagger==0.0.1
lxml==4.9.2
lz4==4.3.2
makefun==1.15.1
Markdown==3.4.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.6.2
mdit-py-plugins==0.3.3
mdurl==0.1.2
missingno==0.5.1
mock==5.0.2
monotonic==1.6
msg-parser==1.2.0
multidict==6.0.4
multimethod==1.9.1
murmurhash==1.0.9
mypy==1.2.0
mypy-extensions==1.0.0
mysql-client==0.0.1
mysql-connector-python==8.0.31
networkx==3.0
newspaper3k==0.2.8
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
oauthlib==3.2.2
olefile==0.46
openai==0.27.3
openapi-schema-pydantic==1.2.4
opencv-python==4.7.0.72
openpyxl==3.1.2
orjson==3.9.1
oscrypto==1.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandas-profiling==3.2.0
passlib==1.7.4
pathy==0.10.1
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.8.0
phik==0.12.3
Pillow==9.4.0
platformdirs==3.0.0
portalocker==2.7.0
preshed==3.0.8
proto-plus==1.22.2
protobuf==4.23.2
psycopg2-binary==2.9.6
pyarrow==10.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycountry==22.3.5
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.7
pydantic_core==2.1.2
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.7.0
pyOpenSSL==23.1.1
pypandoc==1.11
pyparsing==3.0.9
pypdf==3.9.1
PyPDF2==3.0.1
pyrsistent==0.19.3
PySocks==1.7.1
pyspellchecker==0.7.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.1
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
python3-openid==3.2.0
pyttsx3==2.90
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
pyzstd==0.15.6
referencing==0.29.1
regex==2022.10.31
replicate==0.4.0
reportlab==3.6.12
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.0.1
rpds-py==0.8.10
rsa==4.9
s3transfer==0.6.0
sacrebleu==2.3.1
scikit-learn==1.2.0
scipy==1.10.0
seaborn==0.12.2
selenium==4.10.0
semantic-version==2.10.0
setuptools-scm==7.1.0
sgmllib3k==1.0.0
silpa-common==0.3
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
snowflake-connector-python==3.0.2
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sortedcontainers==2.4.0
soundex==1.1.3
soupsieve==2.4
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.9.0
spellchecker==0.4
SQLAlchemy==2.0.16
sqlparse==0.4.3
srsly==2.4.5
starlette==0.26.1
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
termcolor==2.2.0
thinc==8.1.7
threadpoolctl==3.1.0
tiktoken==0.3.2
tinysegmenter==0.3
tldextract==3.4.1
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.1
trio==0.22.1
trio-websocket==0.10.3
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.7
uc-micro-py==1.0.2
unstructured==0.7.12
urllib3==1.26.14
uvicorn==0.22.0
virtualenv==20.20.0
visions==0.7.4
Wand==0.6.11
wasabi==1.1.1
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.2.3
wrapt==1.14.1
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.8.2
